Add doc comments to InputStream type script definition
