+++ js/src/Ice/EndpointFactoryManager.js
@@ -6,7 +6,8 @@ import { ParseException } from "./LocalExceptions.js";
 import { StringUtil } from "./StringUtil.js";
 import { OpaqueEndpointI } from "./OpaqueEndpoint.js";
 import { Protocol } from "./Protocol.js";
-import { OutputStream, InputStream } from "./Stream.js";
+import { OutputStream } from "./OutputStream.js";
+import { InputStream } from "./InputStream.js";
 import { Debug } from "./Debug.js";
 
 export class EndpointFactoryManager {
