+++ js/src/Ice/SliceType.js
@@ -1,0 +1,9 @@
+//
+// Copyright (c) ZeroC, Inc. All rights reserved.
+//
+
+export const SliceType = {
+    NoSlice: 0,
+    ValueSlice: 1,
+    ExceptionSlice: 2,
+};
