+++ js/src/Ice/index.d.ts
@@ -17,7 +17,9 @@
 /// <reference path="./Endpoint.d.ts" />
 /// <reference path="./Properties.d.ts" />
 /// <reference path="./BuiltinSequences.d.ts" />
-/// <reference path="./Stream.d.ts" />
+/// <reference path="./InputStream.d.ts" />
+/// <reference path="./OutputStream.d.ts" />
+/// <reference path="./StreamHelpers.d.ts" />
 /// <reference path="./OperationMode.d.ts" />
 /// <reference path="./MapUtil.d.ts" />
 /// <reference path="./EndpointSelectionType.d.ts" />
